rename js file
