+++ src/site/layout/javascript/artifacts.rs
